Add a prototype implementation of CourtListener as a legal document source
